+++ CHANGELOG.md
@@ -23,3 +23,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixing issue with AddsOrgUnitsAndGroups when OUs contain other non-word characters.
 - Added MmaAgent to configure Microsoft Monitoring Agent.
 - Added AddsServicePrincipalNames to configure SPNs.
+- Disabling RebootNodeIfNeeded when LCM is on Monitor mode.